requirements.txt
pillow
Django==5.1.5
decorator==5.1.1
django-cors-headers==4.7.0
faiss-cpu==1.10.0
groq==0.18.0
langchain==0.3.19
langchain-community==0.3.17
langchain-core==0.3.35
langchain-google-community==2.0.4
langchain-huggingface
langchain-groq==0.2.4
langchain-text-splitters==0.3.6
langchain_community
langchain_core
langchain_google_genai
langgraph==0.2.74
langgraph-checkpoint==2.0.16
langgraph-checkpoint-sqlite==2.0.5
langgraph-sdk==0.1.53
langsmith==0.3.8
oauthlib==3.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1